refactor: unify route guards
