+++ frontend/src/routes/index.tsx
@@ -1,7 +1,6 @@
 import { Routes, Route, Navigate } from "react-router-dom";
 import PageContainer from "../components/layout/PageContainer";
-import PrivateRoute from "../components/auth/PrivateRoute";
-import ProtectedRoute from "../components/auth/ProtectedRoute";
+import AuthRoute from "../components/auth/AuthRoute";
 import { UserRole } from "../types/global";
 import ApplicationLayout from "./calls/apply/ApplicationLayout";
 import Step1_CallInfo from "./calls/apply/Step1_CallInfo";
@@ -30,17 +29,17 @@ export default function AppRoutes() {
       <Route path="/" element={<PageContainer />}>
         <Route index element={<CallsPage />} />
         <Route path="about" element={<AboutPage />} />
-        <Route element={<PrivateRoute roles={[UserRole.admin, UserRole.super_admin]} />}>
+        <Route element={<AuthRoute roles={[UserRole.admin, UserRole.super_admin]} />}>
           <Route path="dashboard" element={<DashboardPage />} />
           <Route path="calls/manage" element={<CallManagementPage />} />
           <Route path="calls/:callId/applications" element={<CallApplicationsPage />} />
         </Route>
         <Route path="calls/:callId" element={<CallDetailPage />} />
         <Route path="calls/:callId/preview" element={<CallPreviewPage />} />
-        <Route element={<ProtectedRoute roles={[UserRole.applicant]} />}>
+        <Route element={<AuthRoute roles={[UserRole.applicant]} />}>
           <Route path="my-applications" element={<MyApplicationsPage />} />
         </Route>
-        <Route element={<ProtectedRoute />}>
+        <Route element={<AuthRoute />}>
           <Route path="calls/:callId/apply" element={<ApplicationLayout />}>
             <Route index element={<Navigate to="step1" replace />} />
             <Route path="step1" element={<Step1_CallInfo />} />
